feat: add script to list rrm_src directory structure

+++ rrm_src/perception/vsa-utils.ts
@@ -109,9 +109,9 @@ export const VSAUtils = {
   },
 
   // 🧠 ENCODER SUPERPOSISI
-  encodeGrid: (grid: number[][], lens: LensType = 'HOLISTIC'): Float32Array => {
+  encodeGrid: (grid: number[][], lens: LensType = 'HOLISTIC'): Float64Array => {
     if (lens === 'COLOR_MAP') {
-        const pixelVecs: Float32Array[] = [];
+        const pixelVecs: Float64Array[] = [];
         for (let y = 0; y < grid.length; y++) {
             for (let x = 0; x < grid[y].length; x++) {
                 const colorVec = ARC_COLORS[grid[y][x]];
@@ -129,7 +129,7 @@ export const VSAUtils = {
     // Jika layar kosong, kembalikan vektor identitas/noise
     if (objects.length === 0) return HoloFFT.create();
 
-    const objectVectors: Float32Array[] = [];
+    const objectVectors: Float64Array[] = [];
 
     // Proses setiap objek secara terpisah menggunakan Lensa Mental
     for (const obj of objects) {
